Incorporate improvements to ga
This commit includes the control adjustments in the main file for varying the mutation. It also incorporates a restructure of how each new generation is generated to ensure that elitism is utilized properly and avg/best fitnesses are updated more reliably. Finally, it includes a minor adjustment in Individual::calcFit(*f) to keep the health of individuals from drifting into the negatives when calculating fitness.